requirements.txt
numpy==1.19.5
Flask==1.1.2
onnxruntime==1.8.1
Pillow==8.3.1
